deletion of user files added

+++ tgbot/data/manage/files/handler.py
@@ -29,14 +29,15 @@ def glob_re(path, regex="", glob_mask="**/*", inverse=False):
 
 def delete_file_by_id(file_path: str, tg_file_id: str):
     file_name_list = glob_re(file_path, regex=f".*{tg_file_id}.*", glob_mask="*")
-    file_name_with_path = file_name_list[0]
-    if os.path.isfile(file_name_with_path):
-        os.remove(file_name_with_path)
-    else:
-        print(f'File in {file_path} is not exist')
+    if file_name_list:
+        file_name_with_path = file_name_list[0]
+        if os.path.isfile(file_name_with_path):
+            os.remove(file_name_with_path)
+        else:
+            print(f'File in {file_path} is not exist')
 
 
-def delete_elder_user_files(num_of_elder: int, num_in_dir_for_user: int, user_id: int, message: Message):
+def delete_elder_user_data(num_of_elder: int, num_in_dir_for_user: int, user_id: int, message: Message):
     """Principle of work:
         Delete num_of_elder from audio dir if number of files in dir > num_in_dir_for_user
         with user id = user_id
@@ -58,8 +59,12 @@ def delete_elder_user_files(num_of_elder: int, num_in_dir_for_user: int, user_id
         while num_of_elder > 0:
             num_of_elder -= 1
             # Get elder audio id from db for user
-            elder_audio_id = sql_handler.get_elder_audio_id(user_id)
+            elder_audio_id = sql_handler.get_elder_audio_tgid(user_id)
+            print('User with elder audio id:', elder_audio_id, 'deleted')
+            # Delete elder input file for user
             delete_file_by_id(audio_input_path, elder_audio_id)
-            # TODO: add id to output file names and add deletion of them
-
+            # Delete elder output file for user
+            delete_file_by_id(audio_output_path, elder_audio_id)
+            # Delete line from table for user
+            sql_handler.delete_from_audio_table(user_id, elder_audio_id)
